refactor: External services: Managers to Services
