+++ app/src/main/java/gateway/services/ServiceWorker.java
@@ -1,11 +1,11 @@
-package gateway.utils;
+package gateway.services;
 
 import capyfile.rmi.IWorkerService;
 import gateway.config.Config;
 import java.rmi.registry.LocateRegistry;
 import java.rmi.registry.Registry;
 
-public class ManagerRMI
+public class ServiceWorker
 {
 	// TODO: Use worker pool
 	public static IWorkerService getServer () throws Exception
